Enable Aserto authorization.
